fix(ci): python-full-coverage — split oversized test files under the 1500-line maintainability limit
test_first_party_code_files_stay_under_line_limit failed because
tests/unit/service/test_host_setup_clients.py (1865 lines) and
tests/unit/cli/test_setup_commands.py (1505 lines) both exceeded
MAX_FIRST_PARTY_FILE_LINES. Split each into a test_*_parts/ directory
matching the repo convention, preserving every test (95 + 55) with no
behavior change.
